job-search/requirements.txt
scrapling[all]
requests
anthropic
python-dotenv
